refactor(Profile) : Modal State 각 컴포넌트 관리하도록 수정 Header 에서 필요한 Modal은 Header에서 정의되도록 리팩토링
